Implement MIDAS FSM V2 Specification & Improved MIDAS Interface & Meta Logging fix & Log Error Detection & Logging Improvements & Log Introspection & Feather duo updates & Pyro I2C bus contention fixes & ADC fixes & GPS update rate changes & HITL & More..

+++ eMMC/src/main.cpp
@@ -68,7 +68,7 @@ void loop() {
     }
 
     if (message == "ls") {
-        size_t tot_size = 0;
+        uint64_t tot_size = 0;
         File root = SD_MMC.open("/");
         if (!root){
             Serial.println("Failed to open directory");
